requirements.txt
roboflow
colorama
torch
torchvision
numpy
matplotlib
pillow